soap-api-python/soap-api-python/requirements.txt
spyne==2.14.0
lxml
mysql-connector-python
